refactor: centralize assets and theme
