+++ src/constants/appConfig.tsx
@@ -1,11 +1,2 @@
 export const APP_NAME = 'Salamtak';
-import logo from '../assets/logo.jpg';
-
-export const LOGO_URL = logo;
 export const BASE_URL = 'http://localhost:8080';
-// src/constants/theme.ts
-export const COLORS = {
-    primary: '#2563eb',   // Tailwind's blue-600
-    secondary: '#1e293b', // slate-800
-    background: '#f8fafc' // slate-50
-};
