+++ src/constants/assets.ts
@@ -1,0 +1,7 @@
+import logo from '../assets/logo.jpg';
+import heroImage from '../assets/doctor2.png';
+import signupBg from '../assets/signup-bg.jpg';
+
+export const LOGO_URL = logo;
+export const HERO_IMAGE = heroImage;
+export const SIGNUP_BG = signupBg;
