+++ src/constants/theme.ts
@@ -1,0 +1,5 @@
+export const COLORS = {
+    primary: '#2563eb',   // Tailwind's blue-600
+    secondary: '#1e293b', // slate-800
+    background: '#f8fafc' // slate-50
+};
